jewerly-shop/backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.5
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
PyJWT==1.7.1
pylint==2.17.4
pytz==2023.3
redis==4.5.5
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
